chore(docs): fix link
I may be incorrect on the best approach, but this link as-is is broken leading to "https://github.com/dlang/DIPs/blob/master/DIPs/accepted/DIP1038.md,%20DIP%201038". This change to LINK2 is based on the README.md in this folder, as well as sibling files like deprecation_delete.dd.